+++ src/events.h
@@ -13,6 +13,7 @@ enum CustomEvent {
   display_alert,
   refresh_data,
   select_location_from_history,
+  select_location_from_list,
   show_about_dialog,
   toggle_current_location_favorite,